Refactor notification handling by replacing `create_notification` with a fluent API using `Notification` and `NotificationManagerCompat`. Remove redundant `channel.py` and improve documentation and examples.
