+++ kvdroid/jclass/androidx/__init__.py
@@ -12,6 +12,9 @@ except JavaException as e:
 from .browser.customtabs import *
 from .core.app import *
 from .core.content import *
+from .core.content.res import *
+from .core.graphics import *
+from .core.graphics.drawable import *
 from .core.view import *
 from .activity import *
 from .appcompat.app import *
